+++ worker/diskmanager/diskmanager_unsupported.go
@@ -5,12 +5,24 @@
 
 package diskmanager
 
-import "github.com/juju/juju/storage"
+import (
+	"github.com/juju/juju/storage"
+	"github.com/juju/juju/version"
+)
 
 var blockDeviceInUse = func(storage.BlockDevice) (bool, error) {
 	panic("not supported")
 }
 
 func listBlockDevices() ([]storage.BlockDevice, error) {
-	panic("not supported")
+	// Return an empty list each time.
+	return nil, nil
+}
+
+func init() {
+	logger.Infof(
+		"block device support has not been implemented for %s",
+		version.Current.OS,
+	)
+	DefaultListBlockDevices = listBlockDevices
 }
